Fix 15 séparation de la définition des axes de l'espace
