+++ src/main/java/org/graphysica/espace2d/forme/Axe.java
@@ -24,21 +24,44 @@ import javafx.beans.property.DoubleProperty;
 import javafx.beans.property.IntegerProperty;
 import javafx.beans.property.SimpleDoubleProperty;
 import javafx.beans.property.SimpleIntegerProperty;
-import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.paint.Color;
 import org.apache.commons.math3.geometry.euclidean.twod.Vector2D;
 import org.graphysica.espace2d.Toile;
 
 /**
+ * Un axe permet de graduer l'espace avec des étiquettes et un sens de
+ * propagation.
  *
  * @author Marc-Antoine
  */
-public class Axe extends Forme {
+public abstract class Axe extends Forme {
+
+    /**
+     * La taille des lignes de tracé de graduations transversales à la flèche
+     * représentant l'axe.
+     */
+    private static final Taille TAILLE_GRADUATION = Taille.de("axegraduation");
 
     /**
      * L'espacement minimum des graduations de l'axe exprimée en pixels.
      */
     private final DoubleProperty espacement = new SimpleDoubleProperty();
+
+    /**
+     * L'ensemble des étiquettes de graduation de l'axe, chacune étant associée
+     * à sa valeur réelle de graduation.
+     */
+    protected final Map<Double, Etiquette> etiquettes = new HashMap<>();
+
+    /**
+     * La flèche représentant cet axe.
+     */
+    protected final Fleche fleche = new Fleche(Vector2D.ZERO, Vector2D.ZERO);
+
+    /**
+     * La taille des caractères des étiquettes de graduation de cet axe.
+     */
+    private final IntegerProperty tailleCaracteres
+            = new SimpleIntegerProperty(14);
 
     /**
      * L'énumération des sens possibles d'un axe.
@@ -49,36 +72,25 @@ public class Axe extends Forme {
     }
 
     /**
-     * Le sens de cet axe.
-     */
-    private final Sens sens;
-
-    /**
-     * L'ensemble des étiquettes de graduation de l'axe, chacune étant associée
-     * à sa valeur réelle de graduation.
-     */
-    private final Map<Double, Etiquette> etiquettes = new HashMap<>();
-
-    /**
-     * La flèche représentant cet axe.
-     */
-    private final Fleche fleche = new Fleche(Vector2D.ZERO, Vector2D.ZERO);
-
-    /**
-     * La taille des caractères des étiquettes de graduation de cet axe.
-     */
-    private final IntegerProperty tailleCaracteres 
-            = new SimpleIntegerProperty(14);
-
-    /**
-     * Construit un axe dont le sens et l'espacement minimale
+     * Construit un nouvel axe dans le sens et d'espacement de graudations
+     * spécifés.
      *
-     * @param sens
-     * @param espacement
+     * @param sens le sens de l'axe.
+     * @param espacement l'espacement minimal entre chacune des graduations de
+     * l'axe.
+     * @return un nouvel axe dans le sens prescrit.
      */
-    public Axe(@NotNull final Sens sens, final double espacement) {
-        this.sens = sens;
-        setEspacement(espacement);
+    public static Axe nouvelAxe(@NotNull final Sens sens,
+            final double espacement) {
+        switch (sens) {
+            case HORIZONTAL:
+                return new AxeHorizontal(espacement);
+            case VERTICAL:
+                return new AxeVertical(espacement);
+            default:
+                throw new UnsupportedOperationException(
+                        "Sens d'axe non supporté: " + sens + ".");
+        }
     }
 
     {
@@ -86,173 +98,25 @@ public class Axe extends Forme {
         proprietesActualisation.add(tailleCaracteres);
     }
 
-    @Override
-    public void dessiner(@NotNull final Toile toile) {
-        switch (sens) {
-            case HORIZONTAL: {
-                final double[] graduationsVerticales = toile
-                        .graduationsVerticales(getEspacement());
-                final double[] abscissesReelles = toile.abscissesReelles(
-                        graduationsVerticales);
-                actualiserEtiquettes(abscissesReelles, formatValeurs(toile));
-                final double positionReelleAxe = positionReelleAxe(toile);
-                fleche.setOrigine(new Vector2D(
-                        toile.abscisseReelle(0), positionReelleAxe));
-                fleche.setArrivee(new Vector2D(
-                        toile.abscisseReelle(toile.getWidth()),
-                        positionReelleAxe));
-                dessinerGraduations(toile, graduationsVerticales,
-                        positionVirtuelleAxe(toile));
-                break;
-            }
-            case VERTICAL: {
-                final double[] graduationsHorizontales = toile
-                        .graduationsHorizontales(getEspacement());
-                final double[] ordonneesReelles = toile.ordonneesReellees(
-                        graduationsHorizontales);
-                actualiserEtiquettes(ordonneesReelles, formatValeurs(toile));
-                final double positionReelleAxe = positionReelleAxe(toile);
-                fleche.setOrigine(new Vector2D(
-                        positionReelleAxe,
-                        toile.ordonneeReelle(toile.getHeight())));
-                fleche.setArrivee(new Vector2D(
-                        positionReelleAxe, toile.ordonneeReelle(0)));
-                dessinerGraduations(toile, graduationsHorizontales,
-                        positionVirtuelleAxe(toile));
-                break;
-            }
-            default: {
-                throw new UnsupportedOperationException(
-                        "Sens d'axe non supporté: " + sens + ".");
-            }
-        }
-        fleche.dessiner(toile);
-        actualiserPositionEtiquettes(toile);
-        etiquettes.values().forEach((etiquette) -> {
-            etiquette.dessiner(toile);
-        });
-    }
-
     /**
-     * Dessine des marques de graduations sur l'axe.
+     * Calcule l'espacement minimal réel entre les graduations de l'axe.
      *
      * @param toile la toile affichant cet axe.
-     * @param valeursVirtuelles les valeurs virtuelles de graduation.
-     * @param positionAxe la position virtuelle de l'axe.
+     * @return la valeur réelle d'espacement minimal des graduations de l'axe.
      */
-    private void dessinerGraduations(@NotNull final Toile toile,
-            @NotNull final double[] valeursVirtuelles,
-            final double positionAxe) {
-        final GraphicsContext contexteGraphique = toile.getGraphicsContext2D();
-        contexteGraphique.setStroke(Color.BLACK);
-        contexteGraphique.setLineWidth(1);
-        switch (sens) {
-            case HORIZONTAL: {
-                for (final double abscisseVirtuelle : valeursVirtuelles) {
-                    contexteGraphique.strokeLine(
-                            abscisseVirtuelle, positionAxe + 3,
-                            abscisseVirtuelle, positionAxe - 3);
-                }
-                break;
-            }
-            case VERTICAL: {
-                for (final double ordonneeVirtuelle : valeursVirtuelles) {
-                    contexteGraphique.strokeLine(
-                            positionAxe + 3, ordonneeVirtuelle,
-                            positionAxe - 3, ordonneeVirtuelle);
-                }
-                break;
-            }
-            default: {
-                throw new UnsupportedOperationException(
-                        "Sens d'axe non supporté: " + sens + ".");
-            }
-        }
-    }
-
-    /**
-     * Calcule la position réelle de l'axe sur la toile. Il peut s'agir de
-     * l'ordonnée ou l'abscisse de l'axe en tant que droite horizontale ou
-     * verticale, le cas échéant.
-     *
-     * @param toile la toile affichant cet axe.
-     * @return l'abscisse ou l'ordonnée de l'axe selon son sens.
-     */
-    private double positionReelleAxe(@NotNull final Toile toile) {
-        switch (sens) {
-            case HORIZONTAL: {
-                final double ordonneeVirtuelleZero = toile.ordonneeVirtuelle(0);
-                if (ordonneeVirtuelleZero < 0) {
-                    return toile.ordonneeReelle(0);
-                } else if (ordonneeVirtuelleZero > toile.getHeight()) {
-                    return toile.ordonneeReelle(toile.getHeight());
-                } else {
-                    return 0;
-                }
-            }
-            case VERTICAL: {
-                final double abscisseVirtuelleZero = toile.abscisseVirtuelle(0);
-                if (abscisseVirtuelleZero < 0) {
-                    return toile.abscisseReelle(0);
-                } else if (abscisseVirtuelleZero > toile.getWidth()) {
-                    return toile.abscisseReelle(toile.getWidth());
-                } else {
-                    return 0;
-                }
-            }
-            default: {
-                throw new UnsupportedOperationException(
-                        "Sens d'axe non supporté: " + sens + ".");
-            }
-        }
-    }
-
-    /**
-     * Calcule la position virtuelle de l'axe sur la toile. Il peut s'agir de
-     * l'ordonnée ou l'abscisse de l'axe en tant que droite horizontale ou
-     * verticale, le cas échéant.
-     *
-     * @param toile la toile affichant cet axe.
-     * @return l'abscisse ou l'ordonnée de l'axe selon son sens.
-     */
-    private double positionVirtuelleAxe(@NotNull final Toile toile) {
-        switch (sens) {
-            case HORIZONTAL: {
-                return toile.ordonneeVirtuelle(positionReelleAxe(toile));
-            }
-            case VERTICAL: {
-                return toile.abscisseVirtuelle(positionReelleAxe(toile));
-            }
-            default: {
-                throw new UnsupportedOperationException(
-                        "Sens d'axe non supporté: " + sens + ".");
-            }
-        }
-    }
+    protected abstract double espacementMinimalReel(@NotNull final Toile toile);
 
     /**
      * Récupère la chaîne de caractères de formattage des valeurs d'étiquette de
-     * l'axe.
+     * l'axe. Le format d'affichage des étiquettes de graduation de l'axe permet
+     * d'arrondir leurs valeurs de telle sorte qu'un nombre minimal de décimales
+     * soient affichées.
      *
      * @param toile la toile sur laquelle l'axe est dessiné.
      * @return le format du texte des étiquettes de graduation.
      */
-    private String formatValeurs(@NotNull final Toile toile) {
-        final double espacementMinimalReel;
-        switch (sens) {
-            case HORIZONTAL:
-                espacementMinimalReel = getEspacement()
-                        / toile.getEchelle().getX();
-                break;
-            case VERTICAL:
-                espacementMinimalReel = getEspacement()
-                        / toile.getEchelle().getY();
-                break;
-            default: {
-                throw new UnsupportedOperationException(
-                        "Sens d'axe non supporté: " + sens + ".");
-            }
-        }
+    protected String formatValeurs(@NotNull final Toile toile) {
+        final double espacementMinimalReel = espacementMinimalReel(toile);
         final int exposant = (int) (Math.log(espacementMinimalReel)
                 / Math.log(Toile.PUISSANCE));
         if (exposant >= 0) {
@@ -267,8 +131,9 @@ public class Axe extends Forme {
      * Actualise la quantité d'étiquettes requises pour le tracé de cet axe.
      *
      * @param valeurs l'ensemble des valeurs représentées par les étiquettes.
+     * @param format le format d'affichage des valeurs d'étiquettes.
      */
-    private void actualiserEtiquettes(@NotNull final double[] valeurs,
+    protected void actualiserEtiquettes(@NotNull final double[] valeurs,
             final String format) {
         retirerEtiquettesObsoletes(valeurs);
         ajouterEtiquettes(valeurs, format);
@@ -282,7 +147,7 @@ public class Axe extends Forme {
      *
      * @param valeurs l'ensemble des valeurs que peuvent prendre les étiquettes.
      */
-    private void retirerEtiquettesObsoletes(final double[] valeurs) {
+    protected void retirerEtiquettesObsoletes(final double[] valeurs) {
         final Iterator<Double> iteration = etiquettes.keySet().iterator();
         while (iteration.hasNext()) {
             final double cle = iteration.next();
@@ -306,7 +171,7 @@ public class Axe extends Forme {
      * @param valeurs l'ensemble des valeurs à afficher sur l'axe.
      * @param format le format des étiquettes.
      */
-    private void ajouterEtiquettes(final double[] valeurs,
+    protected void ajouterEtiquettes(final double[] valeurs,
             final String format) {
         for (final double valeur : valeurs) {
             if (!etiquettes.containsKey(valeur)) {
@@ -317,55 +182,8 @@ public class Axe extends Forme {
         }
     }
 
-    /**
-     * Actualise la position des étiquettes de cet axe.
-     */
-    private void actualiserPositionEtiquettes(@NotNull final Toile toile) {
-        final double positionReelleAxe = positionReelleAxe(toile);
-        final double positionVirtuelleAxe = positionVirtuelleAxe(toile);
-        final Iterator<Map.Entry<Double, Etiquette>> iteration = etiquettes
-                .entrySet().iterator();
-        while (iteration.hasNext()) {
-            final Map.Entry<Double, Etiquette> entree = iteration.next();
-            final double valeur = entree.getKey();
-            final Etiquette etiquette = entree.getValue();
-            switch (sens) {
-                case HORIZONTAL: {
-                    etiquette.setPositionAncrage(
-                            new Vector2D(valeur, positionReelleAxe));
-                    if (positionVirtuelleAxe >= toile.getHeight()
-                            - etiquette.getHauteur()) {
-                        etiquette.setPositionRelative(new Vector2D(
-                                -etiquette.getLargeur() / 2,
-                                toile.getHeight() - positionVirtuelleAxe
-                                - etiquette.getHauteur()));
-                    } else {
-                        etiquette.setPositionRelative(new Vector2D(
-                                -etiquette.getLargeur() / 2, 0));
-                    }
-                    break;
-                }
-                case VERTICAL: {
-                    etiquette.setPositionAncrage(
-                            new Vector2D(positionReelleAxe, valeur));
-                    if (positionVirtuelleAxe >= toile.getWidth()
-                            - etiquette.getLargeur()) {
-                        etiquette.setPositionRelative(new Vector2D(
-                                toile.getWidth() - positionVirtuelleAxe
-                                - etiquette.getLargeur(),
-                                -etiquette.getHauteur() / 2));
-                    } else {
-                        etiquette.setPositionRelative(new Vector2D(
-                                0, -etiquette.getHauteur() / 2));
-                    }
-                    break;
-                }
-                default: {
-                    throw new UnsupportedOperationException(
-                            "Sens d'axe non supporté: " + sens + ".");
-                }
-            }
-        }
+    public int getTailleGraduation() {
+        return TAILLE_GRADUATION.getValue();
     }
 
     public final double getEspacement() {
