+++ src/main/resources/config/taille.properties
@@ -3,3 +3,7 @@ ligne=2
 grille=1
 axe=3
 angle=4
+fleche=4
+teteflechehauteur=6
+teteflechelargeur=4
+axegraduation=3
